Adding testcases
